+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Todos los cambios importantes del proyecto **StopAndGoMX** se documentan en este
 El formato sigue las recomendaciones de **Keep a Changelog**  
 y el versionado está basado en **Semantic Versioning (SemVer)**.
 
+## [1.0.32] - 2026-1-31
+- Managers poder eliminar entrenamientos
+- Bug seleccionar player en documentos
+
 ## [1.0.31] - 2026-1-29
 - Games Detail, Comments, Likes and Photos
 - Player Documents